server: update devices table when listing devices
When creating the list of devices at each start of lavabo-server, it
looks for all available devices and create a dictionary of devices so it
can access devices' commands easily, a row in the devices table and an
empty row in reservations table for this device in the database.

It checks that a row for the device does not already exist in devices
table, if yes, insert a new row and create a new empty row in
reservations table, else abort modification of the database for this
device.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>