lavabo: remove update from available commands
update is no longer a valid method to send to the server and has been
consequently removed from the client.

You could previously call it to ask the list of devices to be updated
because we were thinking to implement lavabo-server as a daemon.
The way we created lavabo is to update the list of available devices
at each start of lavabo-server which is spawned each time an SSH
connection is established (when the client is interacting with the
server). Thus, each time a request is made to the server, the list of
devices is always updated.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>