lavabo: add add-user command
You can now add users to lavabo-server by issuing the add-user request
with a username. This will create a new directory for this user under
the TFTP directory used by LAVA and create a new row in the users table
in the database.

You still need to manually add the user's SSH key (along with the
command given in the README) to lavabo-server authorized_keys to
effectively create a new user.

You need to give lavabo-server sufficient permission to create directory
in the TFTP directory used by LAVA. This can be achieved by doing the
following (assuming lavabo-server is the name of the user running
lavabo-server.py):

 # chgrp lavabo-server TFTP_DIR && chown g+rwx TFTP_DIR

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>