client: send and rename multiple files
This adds the ability to send multiple files to lavabo-server and to
chose their respective name on lavabo-server. For example:

lavabo.py [...] upload file1 -r file

will send the file1 to lavabo-server and rename it file.

You can send multiple files with upload file1 file2 and also rename them
with upload file1 file2 -r foo bar.

If a path to a file is given as argument, the filename is kept. If
renaming is disabled, the filename is kept.

The name of the method is changed from tftp to upload to be more
explicit.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>