server: add check for subdirectory creation
This adds a test during the creation of subdirectory in TFTP directory
to check if the UNIX user who launched lavabo-server has sufficient
permission to create a subdirectory or if the subdirectory is not
already created. If it is, check that the UNIX user who launched
lavabo-server has sufficient permission to write in the subdirectory, if
not, fails to add the user.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>