server: modify response for status requests
This modifies the response sent after a status request. Before, the
sent message was the message received from LAVA server. However, we add
an abstraction layer on the reservation since the board is reserved in
LAVA in name of lavabo-server and is reserved by a user in
lavabo-server. The sent message always gave the name of lavabo-server
instead of the user who reserved the board.

The server now returns the user who effectively reserved the board. It
adds also a reservable state if the board was released but not put
online in LAVA (this behaviour is not implemented yet).

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>