lavabo: automatically add user on first request
When a user had to be created, you had to first add its SSH key and then
make an "add-user" request from a known user to effectively create the
user. Now, you only need to add its SSH key to authorized_keys and the
user will be created the first time this user makes a request. It is
fully transparent for the user, everything has to be done on
server-side.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>