server: catch exceptions raised by xmlrpc
This catches raised XMLRPC Fault exceptions and sends the error code and
explanation to the client. This is helpful when device configuration
file exists in LAVA but the device is not effectively added in LAVA
via the web interface (thus resulting in a 404 error).

This also fixes bypasses of Device methods (XMLRPC API was called
directly outside of Device objects, resulting in code duplication).

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>