lavabo: change default location of lavabo.conf
This commit changes the default location where lavabo searches its
configuration file. Rather than using the same directory as the
directory where the lavabo script is placed, load it from
$HOME/.lavabo.conf.

This is more Unix style, it avoids modifying a file that is under
version control, and potentially opens up for a system-wide
installation of lavabo.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>